Grocery List Program
Asks user what groceries they want, runs through a set list, if the item is in the list, requests quantity and appends to a personal grocery list. If item is not in the list, requests price information from user, then quantity, and appends to personal grocery list. Checks to make sure all input that should be numerical, is.  Prints grocery list with item name, price, quantity and total, as well as a grand total.